refactor: move DatabasePreviewCard
